split home and profile to two separate pages
